+++ src/main/java/com/zerobase/oriticket/domain/members/repository/AdminRepository.java
@@ -1,7 +1,7 @@
-package com.zerobase.oriticket.members.repository;
+package com.zerobase.oriticket.domain.members.repository;
 
-import com.zerobase.global.constants.SignupStatus;
-import com.zerobase.oriticket.members.entity.Admin;
+import com.zerobase.oriticket.domain.members.constants.SignupStatus;
+import com.zerobase.oriticket.domain.members.entity.Admin;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
